requirements.txt
django
pygments
djangorestframework
